+++ tools/rsyslogd.c
@@ -1721,10 +1721,10 @@ mainloop(void)
 			pid_t child;
 			do {
 				child = waitpid(-1, NULL, WNOHANG);
-				DBGPRINTF("rsyslogd: mainloop waitpid (with-no-hang) returned %d\n", child);
+				DBGPRINTF("rsyslogd: mainloop waitpid (with-no-hang) returned %u\n", (unsigned) child);
 				if (child != -1 && child != 0) {
 					errmsg.LogError(0, RS_RET_OK, "Child %d has terminated, reaped "
-						"by main-loop.", child);
+						"by main-loop.", (unsigned) child);
 				}
 			} while(child > 0);
 			bChildDied = 0;
